Logic: rework to reduce unnecessary sql operations
There is a single on_change signal for the logic. It is triggered by a
temperature change, a target temperature change or a mode change. Most
of the time, when the temperature changes, the logic is not impacted so
it's useless to perform actions related to a logic change. Among these
actions is a sql operation which can be very slow on some SD cards.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>