set show label to true for ImageButtons created with an empty string
If an ImageButton/Label is created with an empty string EGT sets the
show_value property to false automatically.
Here, we create ImageButtons for the System and the Fan modes with an
empty string and the text is set later. In this case, the show_label
value must be set to true manually.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>